// shopping-list-app/src/styles/interfaces/iStyleMargin.ts
export default interface IStyleMargin {
  //
  // margin all sides
  //
  m1?: boolean;
  m2?: boolean;
  m3?: boolean;
  m4?: boolean;
  m5?: boolean;
  //
  // margin left
  //
  ml1?: boolean;
  ml2?: boolean;
  ml3?: boolean;
  ml4?: boolean;
  ml5?: boolean;
  //
  // margin right
  //
  mr1?: boolean;
  mr2?: boolean;
  mr3?: boolean;
  mr4?: boolean;
  mr5?: boolean;
  //
  // margin top
  //
  mt1?: boolean;
  mt2?: boolean;
  mt3?: boolean;
  mt4?: boolean;
  mt5?: boolean;
  //
  // margin bottom
  //
  mb1?: boolean;
  mb2?: boolean;
  mb3?: boolean;
  mb4?: boolean;
  mb5?: boolean;
}